# Periodic metrics reporting