Implements https://github.com/istio/ztunnel/issues/926

This PR adds incremental recv/sent bytes metrics recording. This is done by a fork of the tokio copy_buf implementation to increment stats on read/write.

This has two issues currently, which I will open followup issues:
* The `splice` mode is hard to do this way. We will need to fork the library we use or get support upstream. For now, this is disabled
* Currently every r/w operation increments the metrics. This might be a bit excessive. A metric update takes 300ns. A followup could be to only record every N operations (though, ideally, we would also have a way to do it every N seconds or N operations, whichever is smaller...)

For now, I think the benefit here is worth the tradeoffs.

When testing netperf with 2mb payloads and 1byte payloads:

* Before: 10.4Gb/s and 28Mb/s
* After: 10.7Gb/s and 29Mb/s
So roughly in the ballpark. (Presumably,  I did not make things faster, just noise).

Also, pprof shows <1% on metrics reporting