@louiscryan 2024-04-17
Ultimately likely to be better off with thread-local counters and periodic rollup centrally but benefit would be marginal in the grander scheme. Seems fine.